Remove TimeoutFilter
The same behaviour is achieved by using context.WithTimeout
