+++ expire_test.go
@@ -3,33 +3,11 @@ package typhon
 import (
 	"context"
 	"testing"
-	"time"
 
 	"github.com/monzo/terrors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestTimeoutFilter(t *testing.T) {
-	t.Parallel()
-	// A Service which does not time out should be unmolested
-	svc := Service(func(req Request) Response {
-		return Response{}
-	})
-	svc = svc.Filter(TimeoutFilter(time.Second))
-	rsp := svc(NewRequest(nil, "GET", "/", nil))
-	assert.NoError(t, rsp.Error)
-
-	// One which does should time out
-	svc = Service(func(req Request) Response {
-		time.Sleep(50 * time.Millisecond)
-		return Response{}
-	})
-	svc = svc.Filter(TimeoutFilter(10 * time.Millisecond))
-	rsp = svc(NewRequest(nil, "GET", "/", nil))
-	require.Error(t, rsp.Error)
-	assert.True(t, terrors.Wrap(rsp.Error, nil).(*terrors.Error).Matches(terrors.ErrTimeout))
-}
 
 func TestExpirationFilter(t *testing.T) {
 	t.Parallel()
